[REFACTOR] UserDTO → AuthUser 변경 및 auth 도메인으로 이동
- UserDTO → AuthUser (username → email)
- CustomUserDetails 반영

Related to: TICO-376
