+++ backend/pomoro-do/src/main/java/com/tico/pomoro_do/domain/auth/controller/AuthController.java
@@ -3,8 +3,7 @@ package com.tico.pomoro_do.domain.auth.controller;
 import com.tico.pomoro_do.domain.auth.dto.response.TokenResponse;
 import com.tico.pomoro_do.domain.auth.service.AuthService;
 import com.tico.pomoro_do.domain.auth.service.TokenService;
-import com.tico.pomoro_do.global.auth.CustomUserDetails;
-import com.tico.pomoro_do.global.auth.jwt.JWTUtil;
+import com.tico.pomoro_do.domain.auth.security.CustomUserDetails;
 import com.tico.pomoro_do.global.code.ErrorCode;
 import com.tico.pomoro_do.global.code.SuccessCode;
 import com.tico.pomoro_do.global.enums.ProfileImageType;
