+++ backend/pomoro-do/src/main/java/com/tico/pomoro_do/domain/auth/service/CustomUserDetailsService.java
@@ -1,8 +1,9 @@
-package com.tico.pomoro_do.global.auth;
+package com.tico.pomoro_do.domain.auth.service;
 
-import com.tico.pomoro_do.domain.user.dto.UserDTO;
+import com.tico.pomoro_do.domain.auth.dto.AuthUser;
 import com.tico.pomoro_do.domain.user.entity.User;
 import com.tico.pomoro_do.domain.user.repository.UserRepository;
+import com.tico.pomoro_do.domain.auth.security.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -26,31 +27,19 @@ public class CustomUserDetailsService implements UserDetailsService {
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
 
-//        //DB에서 조회
-//        User userData = userRepository.findByUsername(username);
-//
-//        //데이터가 있으면 검증 진행
-//        if (userData != null) {
-//
-//            UserDTO userDTO = UserDTO.builder()
-//                    .username(userData.getUsername())
-//                    .role(String.valueOf(userData.getRole()))
-//                    .build();
-
         // DB에서 조회
         Optional<User> userData = userRepository.findByUsername(username);
-
 
         // 데이터가 있으면 검증 진행
         if (userData.isPresent()) {
             User user = userData.get();
-            UserDTO userDTO = UserDTO.builder()
-                    .username(user.getUsername())
+            AuthUser authUser = AuthUser.builder()
+                    .email(user.getUsername())
                     .role(String.valueOf(user.getRole()))
                     .build();
 
             //UserDetails에 담아서 return하면 AutneticationManager가 검증 함
-            return new CustomUserDetails(userDTO);
+            return new CustomUserDetails(authUser);
         }
 
         return null;
