+++ backend/pomoro-do/src/main/java/com/tico/pomoro_do/domain/user/controller/FollowController.java
@@ -2,7 +2,7 @@ package com.tico.pomoro_do.domain.user.controller;
 
 import com.tico.pomoro_do.domain.user.dto.response.FollowResponse;
 import com.tico.pomoro_do.domain.user.service.FollowService;
-import com.tico.pomoro_do.global.auth.CustomUserDetails;
+import com.tico.pomoro_do.domain.auth.security.CustomUserDetails;
 import com.tico.pomoro_do.global.code.SuccessCode;
 import com.tico.pomoro_do.global.response.SuccessResponseDTO;
 import io.swagger.v3.oas.annotations.Operation;
