--- backend/pomoro-do/src/main/java/com/tico/pomoro_do/domain/user/dto/UserDTO.java
@@ -1,16 +1,0 @@
-package com.tico.pomoro_do.domain.user.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class UserDTO {
-    private String username;
-    private String role;
-
-    @Builder
-    public UserDTO(String username, String role){
-        this.username = username;
-        this.role = role;
-    }
-}
